tweak SuperFastHash algorithm to avoid warning

+++ Library/Sources/Stroika/Foundation/Cryptography/Digest/Algorithm/SuperFastHash.cpp
@@ -63,7 +63,7 @@ void Algorithm::DigesterAlgorithm<Algorithm::SuperFastHash>::Write (const std::b
     size_t len = static_cast<size_t> (end - start);
     Require (len < numeric_limits<uint32_t>::max ());
 
-    fRemainder_ += len;
+    fRemainder_ += static_cast<uint32_t> (len);
     fRemainder_ &= 0x3; // old code just did fRemainder_ = len & 3, but we now get len in bits and peices
     Assert (0 <= fRemainder_ && fRemainder_ <= 3);
